Move tests to NEGATIVE suite
